requirements.txt
torch>=2.0
torchvision
torchaudio
torch-geometric
rdkit-pypi
scikit-learn
pandas
matplotlib
numpy
